move files into src/Input
