+++ src/Button/ButtonPrimary.tsx
@@ -11,7 +11,9 @@ const StyledButton = styled.button`
       ? colors.BATTLESHIP_GREY
       : props.themeObj.colors.primary.dark};
   background-color: ${props =>
-    props.disabled ? colors.DISABLED_GRAY : props.themeObj.colors.primary.light};
+    props.disabled
+      ? colors.DISABLED_GRAY
+      : props.themeObj.colors.primary.light};
   line-height: 36px;
   height: 36px;
   font-size: 14px;
