+++ stories/components/InputNumber.stories.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import { ContentProvider, InputNumber } from "../../../src";
+import { ContentProvider, InputNumber } from "../../src";
 
 import markdown from "./InputNumber.md";
 
 export default {
-  title: "Components/InputNumber"
+  title: "Components/Input"
 };
 
 const contentNodes = {
